requirements.txt
torch==1.9.1
numpy==1.21.2
torchvision==0.10.1
rawpy==0.16.0
Pillow==8.3.2